+++ AMSExplorer/DownloadToLocal.cs
@@ -16,17 +16,9 @@
 
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using Microsoft.WindowsAzure.MediaServices.Client;
-using Excel = Microsoft.Office.Interop.Excel;
-using Microsoft.WindowsAzure.MediaServices.Client.DynamicEncryption;
-using System.Reflection;
 using System.IO;
 using Microsoft.WindowsAPICodePack.Dialogs;